+++ 01 Hardware/01 Lidar/03 Software/01 source/Testen mit USVMonitor/Testen mit USVMonitor/userUnit.h
@@ -65,7 +65,7 @@ typedef struct{
 
 extern uint8_t initUserUnit(usartConfig_t config);
 extern uint8_t usartDataTx(uint8_t* data, uint16_t length);
-extern uint8_t usartDataRx(uint8_t* data, uint8_t length);
+extern uint8_t usartDataRx(uint8_t* data, uint16_t length);
